labeling/requirements.txt
numpy
matplotlib
opencv-python
Pillow
ultralytics